+++ turbo_promptable/lib/spawn/enums/t_spawn_concept.dart
@@ -4,5 +4,4 @@ enum TSpawnConcept {
   role,
   task,
   issue,
-  ;
 }